Update to 1.1

+++ controllers/linkController.js
@@ -1,29 +1,42 @@
-require('dotenv').config()
 const randomString = require('randomstring');
 const mongoose = require('mongoose')
+const validator = require('validator').default
+const qr = require('qrcode')
+const stream = require('stream')
 
 const Link = require('../models/linkModel')
+const User = require('../models/userModel')
 const msg = require('../configs/responseMessages')
-const resFormat = require('../configs/responseFormat')
+const resFormat = require('../configs/responseFormat');
 
 const linkMethods = {
     addNewLink: async (req, res) => {
         try {
-            let userId = mongoose.Types.ObjectId(req.body.userId);
-            let nameLink = req.body.nameLink
-            let longLink = req.body.longLink
-            let shortLink = randomString.generate({
-                length: 8,
-                charset: 'alphanumeric'
-            })
-            let newLink = new Link({
-                owner: userId,
-                nameLink: nameLink,
-                longLink: longLink,
-                shortLink: shortLink
-            })
-            await newLink.save()
-            res.status(201).json(resFormat(true, msg.successCreateLink, ''))
+            let data = async () => {
+                let { userId, nameLink, longLink } = req.body
+                if (!userId || !nameLink || !longLink || !validator.isMongoId(userId) || !validator.isURL(longLink)) {
+                    return { status: 422, data: resFormat(false, msg.failedCreateLink, null) }
+                }
+                userId = mongoose.Types.ObjectId(userId);
+                let shortLink = randomString.generate({
+                    length: 8,
+                    charset: 'alphanumeric'
+                })
+                let user = await User.findById(userId)
+                if (!user) {
+                    return { status: 404, data: resFormat(false, msg.noUserFoundById, null) }
+                }
+                let newLink = new Link({
+                    owner: userId,
+                    nameLink: nameLink,
+                    longLink: longLink,
+                    shortLink: shortLink
+                })
+                let savedLink = await newLink.save()
+                return { status: 201, data: resFormat(true, msg.successCreateLink, savedLink) }
+            }
+            let resp = await data()
+            res.status(resp.status).json(resp.data)
         }
         catch (err) {
             res.status(400).json(resFormat(false, null, err))
@@ -31,36 +44,60 @@ const linkMethods = {
     },
     editLink: async (req, res) => {
         try {
-            let linkId = req.body.linkId;
-            let nameLink = req.body.nameLink
-            let longLink = req.body.longLink
-            await Link.findOneAndUpdate({_id: linkId},{$set:{nameLink: nameLink,longLink: longLink}})
-            res.status(201).json(resFormat(true, msg.successEditLink, ''))
+            let data = async () => {
+                let { linkId, nameLink, longLink } = req.body;
+                if (!linkId || !nameLink || !longLink || !validator.isMongoId(linkId) || !validator.isURL(longLink)) {
+                    return { status: 422, data: resFormat(false, msg.failedEditLink, null) }
+                }
+                let editedLink = await Link.findOneAndUpdate({ _id: linkId }, { $set: { nameLink: nameLink, longLink: longLink } })
+                if (!editedLink) {
+                    return { status: 404, data: resFormat(true, msg.noLinkFoundByLinkId, null) }
+                }
+                return { status: 200, data: resFormat(true, msg.successEditLink, editedLink) }
+            }
+            let resp = await data()
+            res.status(resp.status).json(resp.data)
         }
         catch (err) {
             res.status(400).json(resFormat(false, null, err))
         }
     },
-    deleteLink: async (req,res) =>{
-        try{
-            let linkId = req.body.linkId
-            await Link.findByIdAndDelete(linkId)
-            res.status(201).json(resFormat(true, msg.successDeleteLink, ''))
+    deleteLink: async (req, res) => {
+        try {
+            let data = async () => {
+                let linkId = req.body.linkId
+                if (!linkId || !validator.isMongoId(linkId)) {
+                    return { status: 422, data: resFormat(true, msg.failedDeleteLink, null) }
+                }
+                let deletedLink = await Link.findByIdAndDelete(linkId)
+                if (!deletedLink) {
+                    return { status: 404, data: resFormat(true, msg.noLinkFoundByLinkId, null) }
+                }
+                return { status: 200, data: resFormat(true, msg.successDeleteLink, null) }
+            }
+            let resp = await data()
+            res.status(resp.status).json(resp.data)
         }
-        catch(err){
+        catch (err) {
             res.status(400).json(resFormat(false, null, err))
         }
     },
     getLinkById: async (req, res) => {
         try {
-            let id = req.params.id
-            let data = await Link.findById(id)
-            if (data) {
-                res.status(200).json(resFormat(true, msg.successGetLink, data))
+            let data = async () => {
+                let linkId = req.params.linkId
+                if (!linkId || !validator.isMongoId(linkId)) {
+                    return { status: 422, data: resFormat(false, msg.failedGetLink, null) }
+                }
+                let links = await Link.findById(linkId)
+                if (!links) {
+                    return { status: 404, data: resFormat(false, msg.noLinkFoundByLinkId, null) }
+
+                }
+                return { status: 200, data: resFormat(true, msg.successGetLink, links) }
             }
-            else {
-                res.status(404).json(resFormat(false, msg.noLinkFoundById, data))
-            }
+            let resp = await data()
+            res.status(resp.status).json(resp.data)
         }
         catch (err) {
             res.status(400).json(resFormat(false, null, err))
@@ -68,14 +105,55 @@ const linkMethods = {
     },
     getLinksByUser: async (req, res) => {
         try {
-            let userId = req.params.id
-            let links = await Link.find({ owner: userId }).exec()
-            if (links) {
-                res.status(200).json(resFormat(true, msg.successGetLinks, links))
+            let data = async () => {
+                let userId = req.params.userId
+                if (!userId || !validator.isMongoId(userId)) {
+                    return { status: 422, data: resFormat(false, msg.failedGetLinks, null) }
+                }
+                let links = await Link.find({ owner: userId }).exec()
+                if (links.length === 0) {
+                    return { status: 404, data: resFormat(false, msg.noLinkFoundByUserId, null) }
+                }
+                return { status: 200, data: resFormat(true, msg.successGetLinks, links) }
             }
-            else {
-                res.status(404).json(resFormat(false, msg.noLinkFound, data))
+            let resp = await data()
+            res.status(resp.status).json(resp.data)
+        }
+        catch (err) {
+            res.status(400).json(resFormat(false, null, err))
+        }
+    },
+    redirectToLink: async (req, res) => {
+        try {
+            let data = async () => {
+                let shortLink = req.params.shortLink
+                if (!shortLink) {
+                    return '/'
+                }
+                let links = await Link.findOne({ shortLink: shortLink }).select('longLink')
+                if (!links) {
+                    return '/'
+                }
+                return links.longLink
             }
+            let resp = await data()
+            res.redirect(resp)
+        }
+        catch (err) {
+            res.status(400).json(resFormat(false, null, err))
+        }
+    },
+    getLinkQRCode: async (req, res) => {
+        try {
+            const qrStream = new stream.PassThrough();
+            await qr.toFileStream(qrStream, req.params.shortLink,
+                {
+                    margin: 1,
+                    type: 'png',
+                    width: 720,
+                    errorCorrectionLevel: 'H'
+                });
+            qrStream.pipe(res);
         }
         catch (err) {
             res.status(400).json(resFormat(false, null, err))
